pyproject.toml
[project]
name = "pydanticai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "jinja2>=3.1.5",
    "pydantic-ai>=0.0.18",
    "python-dotenv>=1.0.1",
]
